[REFACTOR] KisApiClient로 분리 #15
