--- src/main/java/com/synergyx/trading/service/kisService/KisCommandService.java
@@ -1,9 +1,0 @@
-package com.synergyx.trading.service.kisService;
-
-public interface KisCommandService {
-    void updateStockDetailsFromKis();
-
-    void updateRoeFromKis();
-
-    void updateRoeFromKisTest(String stockCode);
-}
