+++ src/main/java/com/synergyx/trading/service/kisService/KisPsrUpdateService.java
@@ -1,0 +1,4 @@
+package com.synergyx.trading.service.kisService;
+
+public class KisPsrUpdateService {
+}
